plaid-integration/python/requirements.txt
Flask
plaid_python
python-dotenv
itsdangerous
